requirement.txt
absl-py==0.11.0
args==0.1.0
astunparse==1.6.3
beautifulsoup4==4.9.3
blis==0.7.4
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
chatbot==1.5.2b0
ChatterBot==1.0.8
click==7.1.2
clint==0.5.1
cloudpickle==1.6.0
comtypes==1.1.8
cssselect==1.1.0
cssutils==2.2.0
cymem==2.0.5
dask==2021.2.0
decorator==4.4.2
ecapture==0.1.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
feedparser==6.0.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gTTS==2.2.2
h5py==2.10.0
idna==2.10
importlib-metadata==3.7.0
jaraco.context==4.0.0
Jinja2==2.11.3
joblib==1.0.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==23.0.1
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==1.1.1
mathparse==0.1.2
more-itertools==8.7.0
MouseInfo==0.1.3
multitasking==0.0.9
murmurhash==1.0.5
networkx==2.5
newsapi-python==0.2.6
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
pathy==0.4.0
Pillow==8.1.0
playsound==1.2.2
premailer==3.7.0
preshed==3.0.5
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pydantic==1.7.3
PyGetWindow==0.0.9
pyjokes==0.6.0
PyJWT==1.7.1
pymongo==3.11.3
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.1.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
PyQt5Designer==5.14.1
PyRect==0.1.4
PyScreeze==0.1.26
PySide2==5.15.2
PySimpleGUI==4.38.0
python-dateutil==2.8.1
python-dotenv==0.17.0
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.4.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
regex==2021.8.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.14.2
scipy==1.6.1
selenium==3.141.0
sgmllib3k==1.0.0
shiboken2==5.15.2
six==1.15.0
smart-open==3.0.0
soupsieve==2.2
spacy==3.0.5
spacy-legacy==3.0.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.23
srsly==2.4.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
thinc==8.0.2
toolz==0.11.1
tqdm==4.59.0
twilio==6.53.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wasabi==0.8.2
Werkzeug==1.0.1
wikipedia==1.4.0
winshell==0.6
wolframalpha==5.0.0
wrapt==1.12.1
xmltodict==0.12.0
yagmail==0.14.245
yfinance==0.1.55
zipp==3.4.0
